requirements.txt
attrs==18.1.0
flake8==3.5.0
kafka-python==1.4.2
mccabe==0.6.1
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
pymongo==3.6.1
pytest==3.5.1
six==1.11.0
